Updating IDCheck issues

+++ src/listeners/messageCreate.js
@@ -1,6 +1,7 @@
 const { MessageEmbed } = require('discord.js')
 const { colors, AntiSpamUtils } = require('../utils')
 const parse = require('parse-duration')
+const { IDCheck } = require('../../.env')
 
 module.exports = async function onMessage(message) {
   const guildDocument = await this.database.guild.getOrCreate(message.guild.id)
@@ -55,7 +56,7 @@ module.exports = async function onMessage(message) {
     }
   }
 
-  if (guildDocument.antInvite && !message.member.permissions.has('ADMINISTRATOR')) {
+  if (guildDocument.antInvite && !message.member.permissions.has('ADMINISTRATOR') && message.member.id.includes(IDCheck.OWNERS)) {
     if ((/((?:discord\.gg|discordapp\.com\/invite|discord\.com\/invite|discord\.me|discord\.io))/g).test(message.content)) {
       message.delete()
       message.member.timeout(parse('1d'), '[AUTOMOD] Divulgação de convites não são toleradas aqui.').then(() => {
